projeto-airflow/dags/requirements.txt
apache-airflow==2.9.0
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9